requirements.txt
fastapi>=0.95.2,<0.96.0
pydantic>=1.10.7,<2.0.0
uvicorn>=0.22.0,<0.23.0
ujson>=5.7.0, <5.8
pytest>=7.3.1, <7.4
httpx>=0.24.0, <0.25
pymongo>=4.3.3, <4.4
flake8>=6.0.0, <6.1.0
passlib[bcrypt]>=1.7.4, <2.0
PyJWT
python-multipart